Add register html and svg to api
